+++ ss10/exercise/src/main/java/com/codegym/dto/ProductDto.java
@@ -3,7 +3,7 @@ package com.codegym.dto;
 import java.util.Objects;
 
 public class ProductDto {
-    private int id;
+    private long id;
     private String name;
     private String image;
     private String description;
@@ -12,7 +12,7 @@ public class ProductDto {
     public ProductDto() {
     }
 
-    public ProductDto(int id, String name, String image, String description, long price) {
+    public ProductDto(long id, String name, String image, String description, long price) {
         this.id = id;
         this.name = name;
         this.image = image;
@@ -24,7 +24,7 @@ public class ProductDto {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
